Time: 4 ms (82.94%) | Memory: 42.1 MB (48.74%) - LeetSync

+++ 633-sum-of-square-numbers/sum-of-square-numbers.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean judgeSquareSum(int c) {
         //if (c==0||c==1) return false;
-        long b = (int)Math.sqrt(c);
+        long b = (long)Math.sqrt(c);
         long a=0;
         while(a<=b&&b>=0){
             long sum= (a*a)+(b*b);
